+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.5.3
+
+* test: add tests for zero proximity pager, empty pager, and test current count for all
+  pagers
+* fix(`OffsetPage`): fix `OutOfBoundsException` on an empty first page.
+
 ## 0.5.2
 
 * refactor: `PagerItemInterface` methods now returns itself, instead of